+++ src/object-store/src/parser/parser.hpp
@@ -57,10 +57,17 @@ struct Predicate
         EndsWith,
         Contains
     };
+    
+    enum class OperatorOption
+    {
+        None,
+        CaseInsensitive,
+    };
 
     struct Comparison
     {
         Operator op = Operator::None;
+        OperatorOption option = OperatorOption::None;
         Expression expr[2] = {{Expression::Type::None, ""}, {Expression::Type::None, ""}};
     };