+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.11.1 Release notes (2016-3-29)
+=============================================================
+### Bugfixes
+* Fix for using Android Studio to build app using Realm
+* Fix for sharing Realm between JS and Objective-C/Swift
+
 0.11.0 Release notes (2016-3-24)
 =============================================================
 ### Breaking changes